+++ lib/components/component.dart
@@ -4,6 +4,7 @@ import 'package:flutter/painting.dart';
 import 'package:meta/meta.dart';
 
 import '../effects/effects.dart';
+import '../effects/effects_handler.dart';
 import '../extensions/vector2.dart';
 
 /// This represents a Component for your game.
@@ -12,8 +13,7 @@ import '../extensions/vector2.dart';
 /// Anything that either renders or updates can be added to the list on [BaseGame]. It will deal with calling those methods for you.
 /// Components also have other methods that can help you out if you want to overwrite them.
 abstract class Component {
-  /// The effects that should run on the component
-  final List<ComponentEffect> _effects = [];
+  final EffectsHandler _effectsHandler = EffectsHandler();
 
   /// This method is called periodically by the game engine to request that your component updates itself.
   ///
@@ -22,8 +22,7 @@ abstract class Component {
   /// All components on [BaseGame] are always updated by the same amount. The time each one takes to update adds up to the next update cycle.
   @mustCallSuper
   void update(double dt) {
-    _effects.forEach((e) => e.update(dt));
-    _effects.removeWhere((e) => e.hasFinished());
+    _effectsHandler.update(dt);
   }
 
   /// Renders this component on the provided Canvas [c].
@@ -70,16 +69,19 @@ abstract class Component {
 
   /// Add an effect to the component
   void addEffect(ComponentEffect effect) {
-    _effects.add(effect..initialize(this));
+    _effectsHandler.add(effect, this);
   }
 
   /// Mark an effect for removal on the component
   void removeEffect(ComponentEffect effect) {
-    effect.dispose();
+    _effectsHandler.removeEffect(effect);
   }
 
   /// Remove all effects
   void clearEffects() {
-    _effects.forEach(removeEffect);
+    _effectsHandler.clearEffects();
   }
+
+  /// Get a list of non removed effects
+  List<ComponentEffect> get effects => _effectsHandler.effects;
 }