+++ lib/effects/effects.dart
@@ -1,5 +1,3 @@
-import 'dart:math';
-
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
@@ -19,6 +17,11 @@ abstract class ComponentEffect<T extends Component> {
   bool _isDisposed = false;
   bool get isDisposed => _isDisposed;
 
+  bool _isPaused = false;
+  bool get isPaused => _isPaused;
+  void resume() => _isPaused = false;
+  void pause() => _isPaused = true;
+
   /// If the animation should first follow the initial curve and then follow the
   /// curve backwards
   bool isInfinite;
@@ -27,54 +30,54 @@ abstract class ComponentEffect<T extends Component> {
   final bool _initialIsInfinite;
   final bool _initialIsAlternating;
   double percentage;
-  double travelTime;
+  double curveProgress;
+  double peakTime;
   double currentTime = 0.0;
   double driftTime = 0.0;
   int curveDirection = 1;
+  Curve curve;
 
-  /// If the effect is alternating the travel time is double the normal
-  /// travel time
-  double get totalTravelTime => travelTime * (isAlternating ? 2 : 1);
+  double get iterationTime => peakTime * (isAlternating ? 2 : 1);
 
   ComponentEffect(
     this._initialIsInfinite,
     this._initialIsAlternating, {
     this.isRelative = false,
+    this.curve = Curves.linear,
     this.onComplete,
   }) {
     isInfinite = _initialIsInfinite;
     isAlternating = _initialIsAlternating;
+    curve ??= Curves.linear;
   }
 
+  @mustCallSuper
   void update(double dt) {
-    _updateDriftTime();
     if (isAlternating) {
       curveDirection = isMax() ? -1 : (isMin() ? 1 : curveDirection);
-    } else if (isInfinite && isMax()) {
-      reset();
     }
-    final driftMultiplier = (isAlternating && isMax() ? 2 : 1) * curveDirection;
-    if (!hasFinished()) {
-      currentTime += dt * curveDirection + driftTime * driftMultiplier;
-      percentage = min(1.0, max(0.0, currentTime / travelTime));
-      if (hasFinished()) {
-        onComplete?.call();
+    if (isInfinite) {
+      if ((!isAlternating && isMax()) || (isAlternating && isMin())) {
+        reset();
       }
+    }
+    if (!hasCompleted()) {
+      currentTime += (dt + driftTime) * curveDirection;
+      percentage = (currentTime / peakTime).clamp(0.0, 1.0).toDouble();
+      curveProgress = curve.transform(percentage);
+      _updateDriftTime();
+      currentTime = currentTime.clamp(0.0, peakTime).toDouble();
     }
   }
 
   @mustCallSuper
   void initialize(T component) {
     this.component = component;
-
-    /// You need to set the travelTime during the initialization of the
-    /// extending effect
-    travelTime = null;
   }
 
   void dispose() => _isDisposed = true;
 
-  bool hasFinished() {
+  bool hasCompleted() {
     return (!isInfinite && !isAlternating && isMax()) ||
         (!isInfinite && isAlternating && isMin()) ||
         isDisposed;
@@ -82,26 +85,33 @@ abstract class ComponentEffect<T extends Component> {
 
   bool isMax() => percentage == null ? false : percentage == 1.0;
   bool isMin() => percentage == null ? false : percentage == 0.0;
+  bool isRootEffect() => component?.effects?.contains(this) ?? false;
 
   void reset() {
     _isDisposed = false;
     percentage = null;
     currentTime = 0.0;
     curveDirection = 1;
-    driftTime = 0.0;
     isInfinite = _initialIsInfinite;
     isAlternating = _initialIsAlternating;
+    setComponentToOriginalState();
   }
 
+  // When the time overshoots the max and min it needs to add that time to
+  // whatever is going to happen next, for example an alternation or
+  // following effect in a SequenceEffect.
   void _updateDriftTime() {
     if (isMax()) {
-      driftTime = currentTime - travelTime;
+      driftTime = currentTime - peakTime;
     } else if (isMin()) {
       driftTime = currentTime.abs();
     } else {
       driftTime = 0;
     }
   }
+
+  void setComponentToOriginalState();
+  void setComponentToEndState();
 }
 
 abstract class PositionComponentEffect
@@ -120,11 +130,13 @@ abstract class PositionComponentEffect
     bool initialIsInfinite,
     bool initialIsAlternating, {
     bool isRelative = false,
+    Curve curve,
     void Function() onComplete,
   }) : super(
           initialIsInfinite,
           initialIsAlternating,
           isRelative: isRelative,
+          curve: curve,
           onComplete: onComplete,
         );
 
@@ -133,15 +145,58 @@ abstract class PositionComponentEffect
   void initialize(PositionComponent component) {
     super.initialize(component);
     this.component = component;
-    originalPosition = component.position;
+    originalPosition = component.position.clone();
     originalAngle = component.angle;
-    originalSize = component.size;
+    originalSize = component.size.clone();
 
     /// If these aren't modified by the extending effect it is assumed that the
     /// effect didn't bring the component to another state than the one it
     /// started in
-    endPosition = component.position;
+    endPosition = component.position.clone();
     endAngle = component.angle;
-    endSize = component.size;
+    endSize = component.size.clone();
   }
+
+  @override
+  void setComponentToOriginalState() {
+    if (isRootEffect()) {
+      component?.position?.setFrom(originalPosition);
+      component?.angle = originalAngle;
+      component?.size?.setFrom(endSize);
+    }
+  }
+
+  @override
+  void setComponentToEndState() {
+    if (isRootEffect()) {
+      component?.position?.setFrom(endPosition);
+      component?.angle = endAngle;
+      component?.size?.setFrom(endSize);
+    }
+  }
+}
+
+abstract class SimplePositionComponentEffect extends PositionComponentEffect {
+  double duration;
+  double speed;
+
+  SimplePositionComponentEffect(
+    bool initialIsInfinite,
+    bool initialIsAlternating, {
+    this.duration,
+    this.speed,
+    Curve curve,
+    bool isRelative = false,
+    void Function() onComplete,
+  })  : assert(
+          (duration != null) ^ (speed != null),
+          "Either speed or duration necessary",
+        ),
+        super(
+          initialIsInfinite,
+          initialIsAlternating,
+          isRelative: isRelative,
+          curve: curve,
+          onComplete: onComplete,
+        );
 }