+++ lib/effects/rotate_effect.dart
@@ -3,24 +3,31 @@ import 'package:meta/meta.dart';
 
 import 'effects.dart';
 
-class RotateEffect extends PositionComponentEffect {
-  double radians;
-  double speed;
-  Curve curve;
+class RotateEffect extends SimplePositionComponentEffect {
+  double angle;
   double _startAngle;
   double _delta;
 
+  /// Duration or speed needs to be defined
   RotateEffect({
-    @required this.radians, // As many radians as you want to rotate
-    @required this.speed, // In radians per second
-    this.curve,
+    @required this.angle, // As many radians as you want to rotate
+    double duration, // How long it should take for completion
+    double speed, // The speed of rotation in radians/s
+    Curve curve,
     bool isInfinite = false,
     bool isAlternating = false,
     bool isRelative = false,
     void Function() onComplete,
-  }) : super(
+  })  : assert(
+          (duration != null) ^ (speed != null),
+          "Either speed or duration necessary",
+        ),
+        super(
           isInfinite,
           isAlternating,
+          duration: duration,
+          speed: speed,
+          curve: curve,
           isRelative: isRelative,
           onComplete: onComplete,
         );
@@ -28,18 +35,19 @@ class RotateEffect extends PositionComponentEffect {
   @override
   void initialize(_comp) {
     super.initialize(_comp);
-    if (!isAlternating) {
-      endAngle = _comp.angle + radians;
-    }
     _startAngle = component.angle;
-    _delta = isRelative ? radians : radians - _startAngle;
-    travelTime = (_delta / speed).abs();
+    _delta = isRelative ? angle : angle - _startAngle;
+    if (!isAlternating) {
+      endAngle = _startAngle + _delta;
+    }
+    speed ??= _delta / duration;
+    duration ??= _delta / speed;
+    peakTime = isAlternating ? duration / 2 : duration;
   }
 
   @override
   void update(double dt) {
     super.update(dt);
-    final double progress = curve?.transform(percentage) ?? 1.0;
-    component.angle = _startAngle + _delta * progress;
+    component.angle = _startAngle + _delta * curveProgress;
   }
 }