+++ CHANGELOG.md
@@ -7,6 +7,9 @@
  - Refactor timer class
  - include all changed that are included on 0.28.0
  - Rename game#resize to game#onResize
+ - Test suite for basic effects
+ - Effects duration and test suite for basic effects
+ - Pause and resume for effects
 
 ## 1.0.0-rc1
  - Move all box2d related code and examples to the flame_box2d repo